lge_touch_core: i2c_read: retry a few times before erroring out
Patch from Dennis Rassmann (showp1984) -- I've reduced the number of
retries to 4.

Initial commit message:

--
The i2c devices we are connecting to may not be ready yet, or just
busy. Retry with a 10ms pause in order to still obtain touch data.
This improves overall touch data robustness, but the main thing it
does:
WE CAN FINALLY GET TOUCH DATA IN SUSPEND ALL THE TIME.
--

Signed-off-by: Dennis Rassmann <showp1984@gmail.com>
Signed-off-by: Mikael Berthe <mikael@lilotux.net>